move db browser into own directory
