+++ nems/gui_new/db_browser/db_browser.py
@@ -7,7 +7,7 @@ from PyQt5 import uic
 from PyQt5.QtWidgets import *
 
 from nems import get_setting
-from nems.gui_new.ui_promoted import ListViewListModel
+from nems.gui_new.db_browser.ui_promoted import ListViewListModel
 
 # TEMP ERROR CATCHER
 # Back up the reference to the exceptionhook
