+++ nems/gui_new/db_browser/ui/tab_browser.ui
