+++ nems/gui_new/db_browser/ui/tab_comp.ui
